requirements.txt
Django==4.1
fitz==0.0.1.dev2
llama_index==0.12.1
lxml==4.9.1
Markdown==3.4.4
numpy==1.23.1
openai==1.55.1
pandas==1.5.0
pdfminer==20191125
pdfminer.six==20240706
plotly==5.24.1
pydantic==2.7.0
PyPDF2==3.0.1
python_dateutil==2.8.2
Requests==2.32.3
tiktoken==0.7.0
tqdm==4.66.5
Unidecode==1.2.0
yfinance==0.2.42
django-q
django_extensions
gunicorn
python-decouple
psycopg2-binary
dj_database_url
djangorestframework==3.14.0
drf-nested-routers==0.93.4
django-filter==23.5
schedule
pytz